Feature/fix

+++ WebApplication1/Controllers/CategoryController.cs
@@ -6,6 +6,7 @@ using OrderRestaurant.DTO.CategoryDTO;
 using OrderRestaurant.Helpers;
 using OrderRestaurant.Model;
 using OrderRestaurant.Service;
+using System.Security.Claims;
 using static Azure.Core.HttpHeader;
 
 namespace OrderRestaurant.Controllers
@@ -18,11 +19,14 @@ namespace OrderRestaurant.Controllers
         private readonly ApplicationDBContext _context;
         private readonly ICategory _category;
         private readonly ICommon<CategoryModel> _common;
-        public CategoryController(ApplicationDBContext context, ICategory category, ICommon<CategoryModel> common)
+        private readonly IPermission _permissionRepository;
+        private const string TYPE_Category = "Category";
+        public CategoryController(ApplicationDBContext context, ICategory category, ICommon<CategoryModel> common, IPermission permissionRepository)
         {
             _context = context;
             _category = category;
             _common = common;
+            _permissionRepository = permissionRepository;
         }
 
         [HttpGet("get-search-page")]
@@ -92,34 +96,88 @@ namespace OrderRestaurant.Controllers
        
         public async Task<IActionResult> UpdateCategory([FromRoute] int id, [FromBody] UpdateCategoryDTO updateCategory)
         {
-            var model = await _category.UpdateCategory(id, updateCategory);
-            if (model == null)
+            if (!ModelState.IsValid)
             {
-                return BadRequest();
+                return BadRequest(ModelState);
             }
+            try
+            {
+                var roleName = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+                if (roleName == null)
+                {
+                    return BadRequest("Ko co rolename");
+                }
 
-            return Ok(model.ToCategoryDto());
+                if (!_permissionRepository.CheckPermission(roleName, Constants.Put, TYPE_Category))
+                    return Unauthorized();
+                var model = await _category.UpdateCategory(id, updateCategory);
+                if (model == null)
+                {
+                    return BadRequest();
+                }
 
+                return Ok(model.ToCategoryDto());
+            }
+            catch (Exception ex) {
+                return StatusCode(500, $"Lỗi: {ex.Message}");
+            }
         }
 
         [HttpPost]
         public async Task<IActionResult> Create([FromBody] CreateCategoryDTO category)
         {
-            var model = category.ToCategoryFromCreate();
-            await _category.CreateCategory(model);
-            return CreatedAtAction(nameof(GetCategoryById), new { id = model.CategoryId }, model.ToCategoryDto());
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+            try
+            {
+                var roleName = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+                if (roleName == null)
+                {
+                    return BadRequest("Ko co rolename");
+                }
+
+                if (!_permissionRepository.CheckPermission(roleName, Constants.Post, TYPE_Category))
+                    return Unauthorized();
+                var model = category.ToCategoryFromCreate();
+                await _category.CreateCategory(model);
+                return CreatedAtAction(nameof(GetCategoryById), new { id = model.CategoryId }, model.ToCategoryDto());
+            }catch(Exception ex)
+            {
+                return StatusCode(500, $"Lỗi: {ex.Message}");
+            }
         }
 
         [HttpDelete]
         [Route("{categoryid}")]
         public async Task<IActionResult> DeleteCategory([FromRoute]int categoryid)
         {
-            var model = await _category.DeleteCategory(categoryid);
-            if (model == null)
+            if (!ModelState.IsValid)
             {
-                return NotFound();
+                return BadRequest(ModelState);
             }
-            return NoContent();
+            try
+            {
+                var roleName = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+                if (roleName == null)
+                {
+                    return BadRequest("Ko co rolename");
+                }
+
+                if (!_permissionRepository.CheckPermission(roleName, Constants.Delete, TYPE_Category))
+                    return Unauthorized();
+                var model = await _category.DeleteCategory(categoryid);
+                if (model == null)
+                {
+                    return NotFound();
+                }
+                return NoContent();
+            }
+            catch (Exception ex)
+            {
+                return StatusCode(500, $"Lỗi: {ex.Message}");
+            }
         }
     }
 }
